move stuff outside core into features
