+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/guilds/extensions/Guild.kt
@@ -1,11 +1,14 @@
-package com.mineinabyss.mineinabyss.extensions
+package com.mineinabyss.guilds
 
+import com.mineinabyss.guilds.database.*
+import com.mineinabyss.helpers.MessageQueue
 import com.mineinabyss.idofront.messaging.broadcastVal
 import com.mineinabyss.idofront.messaging.error
 import com.mineinabyss.idofront.messaging.success
 import com.mineinabyss.mineinabyss.core.AbyssContext
 import com.mineinabyss.mineinabyss.core.MIAConfig
-import com.mineinabyss.mineinabyss.data.*
+import com.mineinabyss.mineinabyss.extensions.getGuildName
+import com.mineinabyss.mineinabyss.extensions.getGuildRank
 import org.bukkit.Bukkit
 import org.bukkit.ChatColor
 import org.bukkit.OfflinePlayer
