+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/guilds/menus/GuildInviteScreen.kt
@@ -1,6 +1,7 @@
 package com.mineinabyss.guilds.menus
 
 import androidx.compose.runtime.Composable
+import com.mineinabyss.guilds.database.GuildJoinType
 import com.mineinabyss.guiy.components.Spacer
 import com.mineinabyss.guiy.layout.Row
 import com.mineinabyss.guiy.modifiers.Modifier
@@ -10,7 +11,6 @@ import com.mineinabyss.guiy.modifiers.size
 import com.mineinabyss.helpers.TitleItem
 import com.mineinabyss.helpers.ui.composables.Button
 import com.mineinabyss.idofront.messaging.error
-import com.mineinabyss.mineinabyss.data.GuildJoinType
 import com.mineinabyss.mineinabyss.extensions.*
 import org.bukkit.ChatColor.*
 import org.bukkit.OfflinePlayer
