+++ mineinabyss-features/src/main/kotlin/com/mineinabyss/guilds/menus/GuildJoinRequestScreen.kt
@@ -1,6 +1,7 @@
 package com.mineinabyss.guilds.menus
 
 import androidx.compose.runtime.Composable
+import com.mineinabyss.guilds.database.GuildJoinType
 import com.mineinabyss.guiy.modifiers.Modifier
 import com.mineinabyss.guiy.modifiers.at
 import com.mineinabyss.guiy.modifiers.clickable
@@ -8,7 +9,6 @@ import com.mineinabyss.guiy.modifiers.size
 import com.mineinabyss.helpers.TitleItem
 import com.mineinabyss.helpers.ui.composables.Button
 import com.mineinabyss.idofront.messaging.error
-import com.mineinabyss.mineinabyss.data.GuildJoinType
 import com.mineinabyss.mineinabyss.extensions.addMemberToGuild
 import com.mineinabyss.mineinabyss.extensions.getGuildJoinType
 import com.mineinabyss.mineinabyss.extensions.getGuildMemberCount
