+++ src/main/java/com/mineinabyss/plugin/MineInAbyssPluginImpl.kt
@@ -4,16 +4,16 @@ import com.mineinabyss.geary.minecraft.dsl.GearyAddon
 import com.mineinabyss.geary.minecraft.dsl.GearyLoadPhase
 import com.mineinabyss.geary.minecraft.dsl.gearyAddon
 import com.mineinabyss.geary.minecraft.store.PrefabNamespaceMigrations
+import com.mineinabyss.guilds.database.GuildJoinQueue
+import com.mineinabyss.guilds.database.Guilds
+import com.mineinabyss.guilds.database.Players
+import com.mineinabyss.helpers.MessageQueue
 import com.mineinabyss.idofront.commands.Command
 import com.mineinabyss.idofront.commands.execution.IdofrontCommandExecutor
 import com.mineinabyss.idofront.platforms.IdofrontPlatforms
 import com.mineinabyss.idofront.plugin.getServiceOrNull
 import com.mineinabyss.idofront.plugin.registerService
 import com.mineinabyss.mineinabyss.core.*
-import com.mineinabyss.mineinabyss.data.GuildJoinQueue
-import com.mineinabyss.mineinabyss.data.Guilds
-import com.mineinabyss.mineinabyss.data.MessageQueue
-import com.mineinabyss.mineinabyss.data.Players
 import net.milkbowl.vault.economy.Economy
 import org.bukkit.command.CommandSender
 import org.bukkit.command.TabCompleter
